requirements.txt
fastapi
uvicorn
python-dotenv
cohere